Restore password or invitation token in the same form
