+++ app/assets/javascripts/module-sessions.js
@@ -3,22 +3,13 @@ $(document).on('turbolinks:load', function() {
 	$('.show_login_box').click(function(e) {
 		e.preventDefault();
 		$('.password_recovery_box').addClass("hidden");
-		$('.confirmation_request_box').addClass("hidden");
 		$('.login_box').removeClass('hidden').velocity("transition.slideLeftIn");
 	});
 
 	$('.show_password_recovery_box').click(function(e) {
 		e.preventDefault();
 		$('.login_box').addClass("hidden");
-		$('.confirmation_request_box').addClass("hidden");
 		$('.password_recovery_box').removeClass('hidden').velocity("transition.slideLeftIn");
-	});
-
-	$('.show_confirmation_request_box').click(function(e) {
-		e.preventDefault();
-		$('.login_box').addClass("hidden");
-		$('.password_recovery_box').addClass("hidden");
-		$('.confirmation_request_box').removeClass('hidden').velocity("transition.slideLeftIn");
 	});
 
 });
