+++ app/views/user/sessions/_session_form.html.erb
@@ -14,5 +14,4 @@
   <%= f.submit t(".submit"), class: "button" %>
 
   <%= link_to t(".password_section"), "#", class: "show_password_recovery_box" %>
-  <%= link_to t(".confirmation_request_section"), "#", class: "show_confirmation_request_box" %>
 <% end %>
