+++ app/views/user/sessions/new.html.erb
@@ -20,14 +20,6 @@
 
       </div>
 
-      <div class="confirmation_request_box hidden">
-
-        <h2><%= t(".confirmation_request_title") %></h2>
-
-        <%= render "confirmation_request_form" %>
-
-      </div>
-
     </div>
 
     <div class="pure-u-md-2-24"></div>
