--- config/locales/user/controllers/confirmation_requests/en.yml
@@ -1,7 +1,0 @@
----
-en:
-  user:
-    confirmation_requests:
-      create:
-        error: The email address specified doesn't seem to be valid
-        success: Please check your inbox to get instructions
