+++ config/locales/user/views/sessions/ca.yml
@@ -2,11 +2,7 @@
 ca:
   user:
     sessions:
-      confirmation_request_form:
-        session_section: Ja has confirmat le teu compte? Fes login
-        submit: Sol·licitar
       new:
-        confirmation_request_title: Reenviar instruccions de confirmació
         password_title: 'No recordes la teva contrasenya? T''ajudem:'
         registration_title: 'Nou por aquí? Crea un compte gratis:'
         session_title: Ja tens compte? Fes login
@@ -16,6 +12,5 @@ ca:
       registration_form:
         submit: Anem
       session_form:
-        confirmation_request_section: Instruccions de confirmació?
         password_section: Contrasenya?
         submit: Conectar-se
