+++ config/locales/user/views/sessions/en.yml
@@ -2,11 +2,7 @@
 en:
   user:
     sessions:
-      confirmation_request_form:
-        session_section: Already confirmed? Log in here
-        submit: Request
       new:
-        confirmation_request_title: Send confirmation instructions
         password_title: 'Can''t remember your password? We help you out:'
         registration_title: 'New here? Create your account:'
         session_title: Are you already registered? Log in
@@ -16,6 +12,5 @@ en:
       registration_form:
         submit: Let's go
       session_form:
-        confirmation_request_section: Confirmation instructions?
         password_section: Password?
         submit: Log in
