+++ config/routes.rb
@@ -77,7 +77,6 @@ Rails.application.routes.draw do
       resource :sessions, only: [:new, :create, :destroy]
       resource :registrations, only: [:create]
       resource :confirmations, only: [:new, :create]
-      resource :confirmation_requests, only: [:create]
       resource :passwords, only: [:create, :edit, :update]
       resource :census_verifications, only: [:show, :new, :create], path: :verifications
       resource :settings, only: [:show, :update]
